Jolin

Documentation for Jolin.

Jolin.authenticate_aws

Jolin.jolin_token

Jolin.authenticate_aws — Function

authenticate_aws(role_arn; audience="")

Assume role via web identity. How to define such a role can be found here https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-role.html#cli-configure-role-oidc

CAUTION: Please note that the semicolon is really important! @authorize_aws(role_arn, audience="myaudience") won't work as of now.

Jolin.jolin_token — Function

jolin_token()
jolin_token("exampleaudience")

Creates a JSON Web Token which can be used for authentication at common cloud providers.

On cloud.jolin.io the token will be issued and signed by cloud.jolin.io, on Github Actions (used for automated tests), a respective github token is returned.